Merge NetBSD trunk 20190424-1845

+++ sys/dev/mii/miivar.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: miivar.h,v 1.66 2019/02/26 05:26:10 msaitoh Exp $	*/
+/*	$NetBSD: miivar.h,v 1.68 2019/04/11 09:14:07 msaitoh Exp $	*/
 
 /*-
  * Copyright (c) 1998, 1999, 2000, 2001 The NetBSD Foundation, Inc.
@@ -72,15 +72,11 @@ struct mii_data {
 	LIST_HEAD(mii_listhead, mii_softc) mii_phys;
 	u_int mii_instance;
 
-	/*
-	 * PHY driver fills this in with active media status.
-	 */
+	/* PHY driver fills this in with active media status. */
 	int mii_media_status;
 	u_int mii_media_active;
 
-	/*
-	 * Calls from MII layer into network interface driver.
-	 */
+	/* Calls from MII layer into network interface driver. */
 	mii_readreg_t mii_readreg;
 	mii_writereg_t mii_writereg;
 	mii_statchg_t mii_statchg;
@@ -156,7 +152,7 @@ typedef struct mii_softc mii_softc_t;
 #define	MIIF_IS_HPNA	0x0200		/* is a HomePNA device */
 #define	MIIF_FORCEANEG	0x0400		/* force auto-negotiation */
 
-#define	MIIF_INHERIT_MASK	(MIIF_NOISOLATE|MIIF_NOLOOP|MIIF_AUTOTSLEEP)
+#define	MIIF_INHERIT_MASK (MIIF_NOISOLATE | MIIF_NOLOOP | MIIF_AUTOTSLEEP)
 
 /*
  * Special `locators' passed to mii_attach().  If one of these is not
@@ -196,9 +192,9 @@ struct mii_phydesc {
  * An array of these structures map MII media types to BMCR/ANAR settings.
  */
 struct mii_media {
-	int	mm_bmcr;		/* BMCR settings for this media */
-	int	mm_anar;		/* ANAR settings for this media */
-	int	mm_gtcr;		/* 100base-T2 or 1000base-T CR */
+	uint16_t mm_bmcr;		/* BMCR settings for this media */
+	uint16_t mm_anar;		/* ANAR settings for this media */
+	uint16_t mm_gtcr;		/* 100base-T2 or 1000base-T CR */
 };
 
 #define	MII_MEDIA_NONE		0
@@ -291,9 +287,9 @@ int	mii_mediachg(struct mii_data *);
 void	mii_tick(struct mii_data *);
 void	mii_pollstat(struct mii_data *);
 void	mii_down(struct mii_data *);
-int	mii_anar(int);
+uint16_t mii_anar(int);
 
-int mii_ifmedia_change(struct mii_data *);
+int	mii_ifmedia_change(struct mii_data *);
 
 int	mii_phy_activate(device_t, enum devact);
 int	mii_phy_detach(device_t, int);
@@ -317,7 +313,7 @@ u_int	mii_phy_flowstatus(struct mii_softc *);
 
 void	ukphy_status(struct mii_softc *);
 
-u_int	mii_oui(u_int, u_int);
+u_int	mii_oui(uint16_t, uint16_t);
 #define	MII_OUI(id1, id2)	mii_oui(id1, id2)
 #define	MII_MODEL(id2)		(((id2) & IDR2_MODEL) >> 4)
 #define	MII_REV(id2)		((id2) & IDR2_REV)
